backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
deepeval==2.4.3
google-generativeai==0.8.3
pydantic==2.10.3
giskard>=2.14.0
litellm>=1.0.0
pandas>=2.0.0
